feat: Allow overwriting status codes from the proxied upstream (#4)
Gives a knob to allow changing the upstream's http status to whatever we
desire. This allows glossing over problems in the upstream.

The original sin is something to do with the collector's exporting being
a little out of wack, but...we can't figure it out in an appropriate
timeframe.

As sssuuuccchhh, we add some logging & metrics around what kind of
responses we're seeing to be a little wiser, and hopefully we can
destroy this at a later date! Yay!

Resolves MIG-8854.